+++ lib/src/ble/ble.dart
@@ -1,0 +1,11 @@
+library ble;
+
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'ble.freezed.dart';
+part 'ble.g.dart';
+
+part 'src/device.dart';
+part 'src/service.dart';
+part 'src/characteristic.dart';
+part 'src/property.dart';